Move color class and create Renderer files
